[Refactor] Reduce the need to supply vertex type by non-draw components.

+++ include/objects/cube.h
@@ -3,12 +3,12 @@
 namespace gamo {
     class Cubes {
     public:
-        static GameObject<VertexP3N3C4>* colored(double size = 1);
+        static GameObject<VertexP3N3C4>* colored(float size = 1);
         static GameObject<VertexP3N3T2>* textured(std::string filename, glm::ivec2 tileCount, glm::vec3 size, std::vector<glm::vec2> indices = { });
 
-        static GameObject<VertexP3N3T2>* mc(double size, std::vector<glm::vec2> indices = { });
-        static GameObject<VertexP3N3T2>* mcGrass(double size = 1);
-        static GameObject<VertexP3N3T2>* mcPumpkin(double size = 1);
-        static GameObject<VertexP3N3T2>* mcTotal(double size = 1);
+        static GameObject<VertexP3N3T2>* mc(float size, std::vector<glm::vec2> indices = { });
+        static GameObject<VertexP3N3T2>* mcGrass(float size = 1);
+        static GameObject<VertexP3N3T2>* mcPumpkin(float size = 1);
+        static GameObject<VertexP3N3T2>* mcTotal(float size = 1);
     };
 }
